Quinto commit. Esta hecho casi todo. Solo faltan detalles. Ya es funcional

+++ src/AppBundle/Controller/MaterialController.php
@@ -6,12 +6,14 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use AppBundle\Entity\Materials;
 
 /**
  * Material controller.
  *
  * @Route("/show/material")
+ * @Security("has_role('ROLE_USER')")
  */
 class MaterialController extends Controller
 {
@@ -20,13 +22,21 @@ class MaterialController extends Controller
      *
      * @Route("/", name="show_material_index")
      * @Method("GET")
+	 * @Security("has_role('ROLE_USER')")
      */
     public function indexAction()
     {
+		$whatRole = $this->get('security.authorization_checker');
+		
 		$user = $this->getUser();// para tomar el usuario loggeado
         $em = $this->getDoctrine()->getManager();
-
-		$materials = $em->getRepository('AppBundle:Materials')->findUserMaterialsOrderedByRef($user);// para buscar los materials del usuario
+		
+		if ($whatRole->isGranted('ROLE_ADMIN')){
+			$materials = $em->getRepository('AppBundle:Materials')->findAll();
+		}else
+		{
+			$materials = $em->getRepository('AppBundle:Materials')->findUserMaterialsOrderedByRef($user);// para buscar los materials del usuario
+		}			
 
         return $this->render('material/index.html.twig', array(
             'materials' => $materials,
@@ -38,21 +48,29 @@ class MaterialController extends Controller
      *
      * @Route("/new", name="show_material_new")
      * @Method({"GET", "POST"})
+	 * @Security("has_role('ROLE_USER')")
      */
     public function newAction(Request $request)
     {
+		$whatRole = $this->get('security.authorization_checker');
+		
 		$user = $this->getUser();// para tomar el usuario loggeado
         $material = new Materials();
-        $form = $this->createForm('AppBundle\Form\MaterialType', $material);
+        $form = $this->createForm('AppBundle\Form\MaterialType', $material, ['role' => $this->getUser()->getRoles()]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-			$material->setUserid($user);//le paso el usuario que crea el material
+			if (!$whatRole->isGranted('ROLE_ADMIN')){
+				$material->setUserid($user);//le paso el usuario que crea el material
+			}
             $em = $this->getDoctrine()->getManager();
             $em->persist($material);
             $em->flush();
+			
+			$this->addFlash('notice', 'Material introducido con éxito!');	
 
-            return $this->redirectToRoute('show_material_show', array('id' => $material->getId()));
+            return $this->redirectToRoute('show_material_new');
+			//return $this->redirectToRoute('show_material_show', array('id' => $material->getId()));
         }
 
         return $this->render('material/new.html.twig', array(
@@ -66,6 +84,7 @@ class MaterialController extends Controller
      *
      * @Route("/{id}", name="show_material_show")
      * @Method("GET")
+	 * @Security("has_role('ROLE_USER')")
      */
     public function showAction(Materials $material)
     {
@@ -82,11 +101,12 @@ class MaterialController extends Controller
      *
      * @Route("/{id}/edit", name="show_material_edit")
      * @Method({"GET", "POST"})
+	 * @Security("has_role('ROLE_USER')")
      */
     public function editAction(Request $request, Materials $material)
     {
         $deleteForm = $this->createDeleteForm($material);
-        $editForm = $this->createForm('AppBundle\Form\MaterialType', $material);
+        $editForm = $this->createForm('AppBundle\Form\MaterialType', $material, ['role' => $this->getUser()->getRoles()]);
         $editForm->handleRequest($request);
 
         if ($editForm->isSubmitted() && $editForm->isValid()) {
@@ -109,18 +129,24 @@ class MaterialController extends Controller
      *
      * @Route("/{id}", name="show_material_delete")
      * @Method("DELETE")
+	 * @Security("has_role('ROLE_USER')")
      */
     public function deleteAction(Request $request, Materials $material)
     {
+		$whatRole = $this->get('security.authorization_checker');
         $form = $this->createDeleteForm($material);
         $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($material);
-            $em->flush();
-        }
-
+		if (($this->getUser() == $material->getUserid()) || $whatRole->isGranted('ROLE_ADMIN')) {
+			if ($form->isSubmitted() && $form->isValid()) {
+				$em = $this->getDoctrine()->getManager();
+				$em->remove($material);
+				$em->flush();
+			}
+			$this->addFlash('notice', 'Borrado correcto!');		
+		} else {
+			$this->addFlash('notice', 'No se puede borrar. No eres el propietario ni administrador.');	
+		}
         return $this->redirectToRoute('show_material_index');
     }
