+++ src/AppBundle/Controller/VideoController.php
@@ -6,6 +6,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use AppBundle\Entity\Videos;
 use AppBundle\Entity\Users; //probar
 use AppBundle\Form\VideoType;
@@ -14,6 +15,7 @@ use AppBundle\Form\VideoType;
  * Videos controller.
  *
  * @Route("/show/video")
+ * @Security("has_role('ROLE_USER')")
  */
 class VideoController extends Controller
 {
@@ -22,14 +24,21 @@ class VideoController extends Controller
      *
      * @Route("/", name="show_video_index")
      * @Method("GET")
+	 * @Security("has_role('ROLE_USER')")
      */
     public function indexAction()
     {
+		$whatRole = $this->get('security.authorization_checker');
+		
 		$user = $this->getUser();// para tomar el usuario loggeado
         $em = $this->getDoctrine()->getManager();
-
-        //$videos = $em->getRepository('AppBundle:Videos')->findAll();
-		$videos = $em->getRepository('AppBundle:Videos')->findUserVideosOrderedByName($user);// para buscar los videos del usuario
+		
+		if ($whatRole->isGranted('ROLE_ADMIN')){
+			$videos = $em->getRepository('AppBundle:Videos')->findAll();
+		}else
+		{
+			$videos = $em->getRepository('AppBundle:Videos')->findUserVideosOrderedByName($user);
+		}		
 
         return $this->render('video/index.html.twig', array(
             'videos' => $videos,
@@ -41,16 +50,21 @@ class VideoController extends Controller
      *
      * @Route("/new", name="show_video_new")
      * @Method({"GET", "POST"})
+	 * @Security("has_role('ROLE_USER')")
      */
     public function newAction(Request $request)
     {
+		$whatRole = $this->get('security.authorization_checker');
+		
 		$user = $this->getUser();// para tomar el usuario loggeado
         $video = new Videos();
-        $form = $this->createForm('AppBundle\Form\VideoType', $video);
+        $form = $this->createForm('AppBundle\Form\VideoType', $video, ['role' => $this->getUser()->getRoles()]);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-			$video->setUserid($user);//le paso el usuario que crea el video
+			if (!$whatRole->isGranted('ROLE_ADMIN')){
+				$video->setUserid($user);//le paso el usuario que crea el video
+			}			
             $em = $this->getDoctrine()->getManager();
             $em->persist($video);
             $em->flush();
@@ -69,6 +83,7 @@ class VideoController extends Controller
      *
      * @Route("/{id}", name="show_video_show")
      * @Method("GET")
+	 * @Security("has_role('ROLE_USER')")
      */
     public function showAction(Videos $video)
     {
@@ -85,11 +100,12 @@ class VideoController extends Controller
      *
      * @Route("/{id}/edit", name="show_video_edit")
      * @Method({"GET", "POST"})
+	 * @Security("has_role('ROLE_USER')")
      */
     public function editAction(Request $request, Videos $video)
     {
         $deleteForm = $this->createDeleteForm($video);
-        $editForm = $this->createForm('AppBundle\Form\VideoType', $video);
+        $editForm = $this->createForm('AppBundle\Form\VideoType', $video, ['role' => $this->getUser()->getRoles()]);
         $editForm->handleRequest($request);
 
         if ($editForm->isSubmitted() && $editForm->isValid()) {
@@ -104,7 +120,7 @@ class VideoController extends Controller
         return $this->render('video/edit.html.twig', array(
             'video' => $video,
             'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
+            'delete_form' => $deleteForm->createView()
         ));
     }
 
@@ -113,19 +129,24 @@ class VideoController extends Controller
      *
      * @Route("/{id}", name="show_video_delete")
      * @Method("DELETE")
+	 * @Security("has_role('ROLE_USER')")
      */
     public function deleteAction(Request $request, Videos $video)
     {
+		$whatRole = $this->get('security.authorization_checker');
         $form = $this->createDeleteForm($video);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($video);
-            $em->flush();
-        }
-
-        return $this->redirectToRoute('show_video_index');
+		$form->handleRequest($request);
+		if (($this->getUser() == $video->getUserid()) || $whatRole->isGranted('ROLE_ADMIN')) {
+			if ($form->isSubmitted() && $form->isValid()) {
+				$em = $this->getDoctrine()->getManager();
+				$em->remove($video);
+				$em->flush();
+			}
+			$this->addFlash('notice', 'Borrado correcto!');			
+        } else {
+			$this->addFlash('notice', 'No se puede borrar. No eres el propietario ni administrador.');	
+		}
+		return $this->redirectToRoute('show_video_index');
     }
 
     /**
